requirements.txt
absl-py==2.1.0
addict==2.4.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.1
astunparse==1.6.3
attrs==24.2.0
av==13.0.0
black==21.4b2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.3.0
-e git+https://github.com/Jeff-sjtu/CrowdPose.git@86fe5e11e570cad02d553db52ec72545578cd1c8#egg=crowdposetools&subdirectory=crowdpose-api/PythonAPI
cycler==0.12.1
Cython==3.0.11
decorator==5.1.1
dill==0.3.8
easydict==1.7
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.16.0
flatbuffers==24.3.25
fonttools==4.53.1
fsspec==2024.9.0
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
google-pasta==0.2.0
grpcio==1.66.1
h5py==3.11.0
hydra-core==1.3.2
idna==3.8
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
iopath==0.1.9
ipython==8.18.1
jax==0.4.30
jaxlib==0.4.30
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json-tricks==3.17.3
keras==3.5.0
kiwisolver==1.4.7
lazy_loader==0.4
libclang==18.1.1
lightgbm==4.5.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mediapipe==0.10.15
ml-dtypes==0.4.1
-e git+https://github.com/open-mmlab/mmcv.git@3c9d327cc78b02569abbaa0e4925a2f7be7c73d9#egg=mmcv
mmcv-full==1.7.2
mmdet==3.3.0
mmengine==0.10.4
mpmath==1.3.0
munkres==1.1.4
mypy-extensions==1.0.0
namex==0.0.8
networkx==3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.68
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
Optimizers==0.1
optree==0.12.1
packaging==24.1
pandas==2.2.2
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.3.2
portalocker==2.10.1
prompt_toolkit==3.0.47
protobuf==4.25.4
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycocotools==2.0
pycparser==2.22
pydot==3.0.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyWavelets==1.6.0
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.8.1
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
sounddevice==0.5.0
stack-data==0.6.3
sympy==1.13.2
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2024.8.30
toml==0.10.2
tomli==2.0.1
torch==2.4.0+cu121
torchaudio==2.4.0+cu121
torchvision==0.19.0+cu121
tqdm==4.66.5
traitlets==5.14.3
triton==3.0.0
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.3.15
ultralytics-thop==2.0.6
urllib3==2.2.2
wcwidth==0.2.13
Werkzeug==3.0.4
wrapt==1.16.0
yacs==0.1.8
yapf==0.40.2
zipp==3.20.1
